+++ centaur/src/main/scala/centaur/api/CentaurCromwellClient.scala
@@ -5,13 +5,14 @@ import java.util.concurrent.Executors
 
 import akka.actor.ActorSystem
 import akka.http.scaladsl.Http
-import akka.http.scaladsl.model.HttpRequest
+import akka.http.scaladsl.model.{HttpRequest, StatusCodes}
 import akka.http.scaladsl.unmarshalling.Unmarshaller.UnsupportedContentTypeException
 import akka.stream.{ActorMaterializer, ActorMaterializerSettings, StreamTcpException}
 import centaur.test.metadata.WorkflowMetadata
 import centaur.test.workflow.Workflow
 import centaur.{CentaurConfig, CromwellManager}
 import cromwell.api.CromwellClient
+import cromwell.api.CromwellClient.UnsuccessfulRequestException
 import cromwell.api.model.{CromwellBackends, SubmittedWorkflow, WorkflowId, WorkflowOutputs, WorkflowStatus}
 
 import scala.concurrent._
@@ -84,6 +85,9 @@ object CentaurCromwellClient {
            _: StreamTcpException |
            _: IOException |
            _: UnsupportedContentTypeException if attempt < awaitMaxAttempts =>
+        Thread.sleep(awaitSleep.toMillis)
+        awaitFutureCompletion(x, timeout, attempt + 1)
+      case unsuccessful: UnsuccessfulRequestException if unsuccessful.httpResponse.status == StatusCodes.NotFound && attempt < awaitMaxAttempts =>
         Thread.sleep(awaitSleep.toMillis)
         awaitFutureCompletion(x, timeout, attempt + 1)
       // see https://github.com/akka/akka-http/issues/768