+++ supportedBackends/google/pipelines/common/src/main/scala/cromwell/backend/google/pipelines/common/PipelinesApiRuntimeAttributes.scala
@@ -6,15 +6,17 @@ import cats.syntax.validated._
 import com.typesafe.config.Config
 import common.validation.ErrorOr._
 import cromwell.backend.google.pipelines.common.GpuResource.GpuType.GpuType
-import cromwell.backend.google.pipelines.common.io.{JesAttachedDisk, JesWorkingDisk}
+import cromwell.backend.google.pipelines.common.io.{PipelinesApiAttachedDisk, PipelinesApiWorkingDisk}
 import cromwell.backend.standard.StandardValidatedRuntimeAttributesBuilder
 import cromwell.backend.validation.{BooleanRuntimeAttributesValidation, _}
 import wom.RuntimeAttributesKeys
 import wom.format.MemorySize
 import wom.types._
 import wom.values._
+import net.ceedubs.ficus.Ficus._
 
 object GpuResource {
+  val DefaultNvidiaDriverVersion = "390.46"
   object GpuType extends Enumeration {
     type GpuType = Value
     val NVIDIATeslaP100 = Value("nvidia-tesla-p100")
@@ -22,7 +24,7 @@ object GpuResource {
   }
 }
 
-final case class GpuResource(gpuType: GpuType, gpuCount: Int)
+final case class GpuResource(gpuType: GpuType, gpuCount: Int, nvidiaDriverVersion: String = GpuResource.DefaultNvidiaDriverVersion)
 
 final case class PipelinesApiRuntimeAttributes(cpu: Int,
                                                gpuResource: Option[GpuResource],
@@ -30,7 +32,7 @@ final case class PipelinesApiRuntimeAttributes(cpu: Int,
                                                preemptible: Int,
                                                bootDiskSize: Int,
                                                memory: MemorySize,
-                                               disks: Seq[JesAttachedDisk],
+                                               disks: Seq[PipelinesApiAttachedDisk],
                                                dockerImage: String,
                                                failOnStderr: Boolean,
                                                continueOnReturnCode: ContinueOnReturnCode,
@@ -54,7 +56,7 @@ object PipelinesApiRuntimeAttributes {
   private val NoAddressDefaultValue = WomBoolean(false)
 
   val DisksKey = "disks"
-  private val DisksDefaultValue = WomString(s"${JesWorkingDisk.Name} 10 SSD")
+  private val DisksDefaultValue = WomString(s"${PipelinesApiWorkingDisk.Name} 10 SSD")
 
   private val MemoryDefaultValue = "2 GB"
 
@@ -74,7 +76,7 @@ object PipelinesApiRuntimeAttributes {
 
   private def continueOnReturnCodeValidation(runtimeConfig: Option[Config]) = ContinueOnReturnCodeValidation.default(runtimeConfig)
 
-  private def disksValidation(runtimeConfig: Option[Config]): RuntimeAttributesValidation[Seq[JesAttachedDisk]] = DisksValidation
+  private def disksValidation(runtimeConfig: Option[Config]): RuntimeAttributesValidation[Seq[PipelinesApiAttachedDisk]] = DisksValidation
     .withDefault(DisksValidation.configDefaultWomValue(runtimeConfig) getOrElse DisksDefaultValue)
 
   private def zonesValidation(runtimeConfig: Option[Config]): RuntimeAttributesValidation[Vector[String]] = ZonesValidation
@@ -126,10 +128,11 @@ object PipelinesApiRuntimeAttributes {
     val cpu: Int = RuntimeAttributesValidation.extract(cpuValidation(runtimeAttrsConfig), validatedRuntimeAttributes)
 
     // GPU
+    lazy val nvidiaDriverVersion = runtimeAttrsConfig.flatMap(_.as[Option[String]]("nvidia-driver-version")).getOrElse(GpuResource.DefaultNvidiaDriverVersion)
     val gpuType: Option[GpuType] = RuntimeAttributesValidation.extractOption(gpuTypeValidation(runtimeAttrsConfig).key, validatedRuntimeAttributes)
     val gpu: Option[Int] = RuntimeAttributesValidation.extractOption(gpuValidation(runtimeAttrsConfig).key, validatedRuntimeAttributes)
     val gpuResource = (gpuType, gpu) match {
-      case (Some(t), Some(g)) => Option(GpuResource(t, g))
+      case (Some(t), Some(g)) => Option(GpuResource(t, g, nvidiaDriverVersion))
       case (Some(_), None) => throw new RuntimeException(s"Please specify how many GPU should be attached to the instance.")
       case (None, Some(_)) => throw new RuntimeException(s"Please specify a GPU type: ${GpuResource.GpuType.values.mkString(", ")}")
       case (None, None) => None
@@ -139,7 +142,7 @@ object PipelinesApiRuntimeAttributes {
     val preemptible: Int = RuntimeAttributesValidation.extract(preemptibleValidation(runtimeAttrsConfig), validatedRuntimeAttributes)
     val bootDiskSize: Int = RuntimeAttributesValidation.extract(bootDiskSizeValidation(runtimeAttrsConfig), validatedRuntimeAttributes)
     val memory: MemorySize = RuntimeAttributesValidation.extract(memoryValidation(runtimeAttrsConfig), validatedRuntimeAttributes)
-    val disks: Seq[JesAttachedDisk] = RuntimeAttributesValidation.extract(disksValidation(runtimeAttrsConfig), validatedRuntimeAttributes)
+    val disks: Seq[PipelinesApiAttachedDisk] = RuntimeAttributesValidation.extract(disksValidation(runtimeAttrsConfig), validatedRuntimeAttributes)
     val docker: String = RuntimeAttributesValidation.extract(dockerValidation, validatedRuntimeAttributes)
     val failOnStderr: Boolean = RuntimeAttributesValidation.extract(failOnStderrValidation(runtimeAttrsConfig), validatedRuntimeAttributes)
     val continueOnReturnCode: ContinueOnReturnCode = RuntimeAttributesValidation.extract(continueOnReturnCodeValidation(runtimeAttrsConfig), validatedRuntimeAttributes)
@@ -176,43 +179,43 @@ object ZonesValidation extends RuntimeAttributesValidation[Vector[String]] {
     s"Expecting $key runtime attribute to be either a whitespace separated String or an Array[String]"
 }
 
-object DisksValidation extends RuntimeAttributesValidation[Seq[JesAttachedDisk]] {
+object DisksValidation extends RuntimeAttributesValidation[Seq[PipelinesApiAttachedDisk]] {
   override def key: String = PipelinesApiRuntimeAttributes.DisksKey
 
   override def coercion: Traversable[WomType] = Set(WomStringType, WomArrayType(WomStringType))
 
-  override protected def validateValue: PartialFunction[WomValue, ErrorOr[Seq[JesAttachedDisk]]] = {
+  override protected def validateValue: PartialFunction[WomValue, ErrorOr[Seq[PipelinesApiAttachedDisk]]] = {
     case WomString(value) => validateLocalDisks(value.split(",\\s*").toSeq)
     case WomArray(womType, values) if womType.memberType == WomStringType =>
       validateLocalDisks(values.map(_.valueString))
   }
 
-  private def validateLocalDisks(disks: Seq[String]): ErrorOr[Seq[JesAttachedDisk]] = {
-    val diskNels: Seq[ErrorOr[JesAttachedDisk]] = disks map validateLocalDisk
-    val sequenced: ErrorOr[Seq[JesAttachedDisk]] = sequenceNels(diskNels)
-    val defaulted: ErrorOr[Seq[JesAttachedDisk]] = addDefault(sequenced)
+  private def validateLocalDisks(disks: Seq[String]): ErrorOr[Seq[PipelinesApiAttachedDisk]] = {
+    val diskNels: Seq[ErrorOr[PipelinesApiAttachedDisk]] = disks map validateLocalDisk
+    val sequenced: ErrorOr[Seq[PipelinesApiAttachedDisk]] = sequenceNels(diskNels)
+    val defaulted: ErrorOr[Seq[PipelinesApiAttachedDisk]] = addDefault(sequenced)
     defaulted
   }
 
-  private def validateLocalDisk(disk: String): ErrorOr[JesAttachedDisk] = {
-    JesAttachedDisk.parse(disk) match {
+  private def validateLocalDisk(disk: String): ErrorOr[PipelinesApiAttachedDisk] = {
+    PipelinesApiAttachedDisk.parse(disk) match {
       case scala.util.Success(attachedDisk) => attachedDisk.validNel
       case scala.util.Failure(ex) => ex.getMessage.invalidNel
     }
   }
 
-  private def sequenceNels(nels: Seq[ErrorOr[JesAttachedDisk]]): ErrorOr[Seq[JesAttachedDisk]] = {
-    val emptyDiskNel: ErrorOr[Vector[JesAttachedDisk]] = Vector.empty[JesAttachedDisk].validNel
-    val disksNel: ErrorOr[Vector[JesAttachedDisk]] = nels.foldLeft(emptyDiskNel) {
+  private def sequenceNels(nels: Seq[ErrorOr[PipelinesApiAttachedDisk]]): ErrorOr[Seq[PipelinesApiAttachedDisk]] = {
+    val emptyDiskNel: ErrorOr[Vector[PipelinesApiAttachedDisk]] = Vector.empty[PipelinesApiAttachedDisk].validNel
+    val disksNel: ErrorOr[Vector[PipelinesApiAttachedDisk]] = nels.foldLeft(emptyDiskNel) {
       (acc, v) => (acc, v) mapN { (a, v) => a :+ v }
     }
     disksNel
   }
 
-  private def addDefault(disksNel: ErrorOr[Seq[JesAttachedDisk]]): ErrorOr[Seq[JesAttachedDisk]] = {
+  private def addDefault(disksNel: ErrorOr[Seq[PipelinesApiAttachedDisk]]): ErrorOr[Seq[PipelinesApiAttachedDisk]] = {
     disksNel map {
-      case disks if disks.exists(_.name == JesWorkingDisk.Name) => disks
-      case disks => disks :+ JesWorkingDisk.Default
+      case disks if disks.exists(_.name == PipelinesApiWorkingDisk.Name) => disks
+      case disks => disks :+ PipelinesApiWorkingDisk.Default
     }
   }