+++ supportedBackends/google/pipelines/common/src/main/scala/cromwell/backend/google/pipelines/common/api/PipelinesApiRequestWorker.scala
@@ -17,7 +17,7 @@ import scala.util.{Failure, Success, Try}
   * Sends batched requests to JES as a worker to the JesApiQueryManager
   */
 class PipelinesApiRequestWorker(val pollingManager: ActorRef, val batchInterval: FiniteDuration, override val serviceRegistryActor: ActorRef)
-                               (implicit val batchHandler: PipelinesApiBatchHandler)
+                               (implicit val batchHandler: PipelinesApiRequestHandler)
   extends Actor with ActorLogging with CromwellInstrumentationActor {
   // The interval to delay between submitting each batch
   log.info("JES batch polling interval is {}", batchInterval)
@@ -53,7 +53,7 @@ class PipelinesApiRequestWorker(val pollingManager: ActorRef, val batchInterval:
   
   // These are separate functions so that the tests can hook in and replace the JES-side stuff
   private[api] def createBatch(): BatchRequest = batch
-  private[api] def runBatch(batch: BatchRequest) = batch.execute()
+  private[api] def runBatch(batch: BatchRequest) = if (batch.size() > 0) batch.execute()
 
   // TODO: FSMify this actor?
   private def interstitialRecombobulation: PartialFunction[Try[List[Try[Unit]]], Unit] = {
@@ -80,14 +80,14 @@ class PipelinesApiRequestWorker(val pollingManager: ActorRef, val batchInterval:
     * Warning: Only use this from inside a receive method.
     */
   private def scheduleCheckForWork(): Unit = {
-    context.system.scheduler.scheduleOnce(batchInterval) { pollingManager ! PipelinesApiRequestManager.RequestJesPollingWork(MaxBatchSize) }
+    context.system.scheduler.scheduleOnce(batchInterval) { pollingManager ! PipelinesApiRequestManager.PipelinesWorkerRequestWork(MaxBatchSize) }
     ()
   }
 }
 
 object PipelinesApiRequestWorker {
   def props(pollingManager: ActorRef, batchInterval: FiniteDuration, serviceRegistryActor: ActorRef)
-              (implicit batchHandler: PipelinesApiBatchHandler) = {
+              (implicit batchHandler: PipelinesApiRequestHandler) = {
     Props(new PipelinesApiRequestWorker(pollingManager, batchInterval, serviceRegistryActor)).withDispatcher(BackendDispatcher)
   }