+++ supportedBackends/google/pipelines/common/src/main/scala/cromwell/backend/google/pipelines/common/api/clients/PipelinesApiRunCreationClient.scala
@@ -3,7 +3,7 @@ package cromwell.backend.google.pipelines.common.api.clients
 import akka.actor.{Actor, ActorLogging, ActorRef}
 import cromwell.backend.google.pipelines.common.PapiInstrumentation
 import cromwell.backend.google.pipelines.common.api.PipelinesApiRequestFactory.CreatePipelineParameters
-import cromwell.backend.google.pipelines.common.api.PipelinesApiRequestManager.{JesApiRunCreationQueryFailed, PAPIApiException}
+import cromwell.backend.google.pipelines.common.api.PipelinesApiRequestManager.{PipelinesApiRunCreationQueryFailed, PAPIApiException}
 import cromwell.backend.google.pipelines.common.api.{PipelinesApiRequestFactory, PipelinesApiRequestManager}
 import cromwell.backend.standard.StandardAsyncJob
 import cromwell.core.WorkflowId
@@ -36,7 +36,7 @@ trait PipelinesApiRunCreationClient { this: Actor with ActorLogging with PapiIns
     case job: StandardAsyncJob =>
       runSuccess()
       completePromise(Success(job))
-    case JesApiRunCreationQueryFailed(_, e) => completePromise(Failure(e))
+    case PipelinesApiRunCreationQueryFailed(_, e) => completePromise(Failure(e))
   }
 
   private def completePromise(job: Try[StandardAsyncJob]) = {
@@ -49,7 +49,7 @@ trait PipelinesApiRunCreationClient { this: Actor with ActorLogging with PapiIns
       case Some(p) =>
         p.future
       case None =>
-        papiApiActor ! PipelinesApiRequestManager.PAPIRunCreationRequest(workflowId, self, requestFactory.makeRunPipelineRequest(createPipelineParameters))
+        papiApiActor ! PipelinesApiRequestManager.PAPIRunCreationRequest(workflowId, self, requestFactory.runRequest(createPipelineParameters))
         val newPromise = Promise[StandardAsyncJob]()
         runCreationClientPromise = Option(newPromise)
         newPromise.future