+++ supportedBackends/google/pipelines/v1alpha2/src/test/scala/cromwell/backend/google/pipelines/v1alpha2/api/PipelinesApiRequestWorkerV1Spec.scala
@@ -8,8 +8,9 @@ import com.google.api.services.genomics.model.Operation
 import cromwell.backend.google.pipelines.common.PipelinesApiTestConfig.jesConfiguration
 import cromwell.backend.google.pipelines.common.api.PipelinesApiRequestManager.PAPIStatusPollRequest
 import cromwell.backend.google.pipelines.common.api.{PipelinesApiRequestManager, PipelinesApiRequestWorkerSpec, TestPipelinesApiBatchHandler, TestPipelinesApiRequestWorker}
+import cromwell.backend.google.pipelines.v1alpha2.api.request.RequestHandler
 
-import scala.concurrent.{Future, Promise}
+import scala.concurrent.{ExecutionContext, Future, Promise}
 import scala.util.{Success, Try}
 
 class PipelinesApiRequestWorkerV1Spec extends PipelinesApiRequestWorkerSpec[Operation] {
@@ -24,7 +25,7 @@ class PipelinesApiRequestWorkerV1Spec extends PipelinesApiRequestWorkerSpec[Oper
 }
 
 class TestPipelinesApiBatchHandlerV1(pollingManager: ActorRef) extends TestPipelinesApiBatchHandler[Operation] {
-  val delegate = new BatchHandler("cromwell-test-app") {
+  val delegate = new RequestHandler("cromwell-test-app") {
     override def interpretOperationStatus(operation: Operation) = {
       mockStatusInterpreter(operation)
     }
@@ -36,7 +37,8 @@ class TestPipelinesApiBatchHandlerV1(pollingManager: ActorRef) extends TestPipel
   }
   
   override def makeBatchRequest = null
-  override def enqueue[T <: PipelinesApiRequestManager.PAPIApiRequest](papiApiRequest: T, batchRequest: BatchRequest, pollingManager: ActorRef) = {
+  override def enqueue[T <: PipelinesApiRequestManager.PAPIApiRequest](papiApiRequest: T, batchRequest: BatchRequest, pollingManager: ActorRef)
+                                                                      (implicit ec: ExecutionContext)= {
     papiApiRequest match {
       case poll: PAPIStatusPollRequest => enqueueStatusPollInBatch(poll, batchRequest)
       case _ => Future.successful(Success(()))